Move LocalizedTextProvider to Sandra.UI.csproj.
